Working solution using aws lambda.
The problem with this solution is that it does not effectively make use of the cache of
previously computed values. There is no guarentee that the same warm lambda will pick up
and handle every request, meaning that in the worst case a lambda will time out trying
to handle fib(n) where n is arbitrarily high. If there is a cold start for large value of N, then
there would be no precomputed cache to inspect and therefore requests would take much longer.

There are two other approaches that I will investigate. I could use one or a combination of both:
1. Use a long lived server. This may be better since there is a guarentee that
the same server will pick up every request (so long as there is only one server, of course).
2. Pre calculate the Fibonacci sequence and store the sequence off to the side - either in memory
or in the database. We could then quickly pull and reference the precomputed sequence on every invocation.
Sorting this precomputed sequence in memory would be a lot quicker and the downside of storing
the sequence in the database is, well, until what value of N do we store, and what happens if we go over it?